Move some of the helper functions into a SlackUtils module
